+++ spring/spring-web-reactive/src/test/java/io/quarkus/ts/spring/web/reactive/reactive/MariaDBConstants.java
@@ -1,0 +1,8 @@
+package io.quarkus.ts.spring.web.reactive.reactive;
+
+public class MariaDBConstants {
+    public static final int PORT = 3306;
+    public static final String START_LOG = "Only MySQL server logs after this point";
+    public static final String IMAGE_103 = "${mariadb.103.image}";
+    public static final String IMAGE_105 = "${mariadb.105.image}";
+}
