+++ spring/spring-web-reactive/src/test/java/io/quarkus/ts/spring/web/reactive/reactive/boostrap/AbstractSpringWebQuteReactiveIT.java
@@ -1,4 +1,4 @@
-package io.quarkus.ts.spring.web.reactive.reactive;
+package io.quarkus.ts.spring.web.reactive.reactive.boostrap;
 
 import org.hamcrest.CoreMatchers;
 import org.junit.jupiter.api.Test;
@@ -7,7 +7,7 @@ import org.springframework.http.HttpStatus;
 import io.quarkus.test.bootstrap.RestService;
 import io.restassured.http.ContentType;
 
-public abstract class AbstractDbReactiveIT {
+public abstract class AbstractSpringWebQuteReactiveIT {
 
     private static final String APP_NAME = "Bootstrap Spring Boot";
