+++ spring/spring-web-reactive/src/test/java/io/quarkus/ts/spring/web/reactive/reactive/openapi/AbstractSpringWebOpenApiReactiveIT.java
@@ -6,50 +6,21 @@ import java.util.stream.Stream;
 
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
-import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvFileSource;
 
-import io.quarkus.test.bootstrap.MariaDbService;
 import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.scenarios.QuarkusScenario;
-import io.quarkus.test.services.Container;
-import io.quarkus.test.services.QuarkusApplication;
-import io.quarkus.ts.spring.web.reactive.reactive.AbstractDbReactiveIT;
 import io.restassured.response.Response;
 
-@QuarkusScenario
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-@EnabledIfSystemProperty(named = "ts.redhat.registry.enabled", matches = "true")
-public class OpenApiReactiveIT extends AbstractDbReactiveIT {
+public abstract class AbstractSpringWebOpenApiReactiveIT {
     private static Response response;
 
-    private static final String API_ROOT = "/api/books";
-
-    static final int MARIADB_PORT = 3306;
-
-    @Container(image = "${mariadb.105.image}", port = MARIADB_PORT, expectedLog = "Only MySQL server logs after this point")
-    static final MariaDbService database = new MariaDbService();
-
-    @QuarkusApplication
-    private static final RestService app = new RestService()
-            .withProperty("quarkus.datasource.username", database.getUser())
-            .withProperty("quarkus.datasource.password", database.getPassword())
-            .withProperty("quarkus.datasource.reactive.url",
-                    () -> "vertx-reactive:" + database.getHost().replace("http", "mysql") + ":" + database.getPort() + "/"
-                            + database.getDatabase());
-
-    @Override
-    public RestService getApp() {
-        return app;
-    }
+    protected abstract RestService getApp();
 
     private void callOpenApiEndpoint() {
-        response = app.given().get("/q/openapi?format=JSON");
+        response = getApp().given().get("/q/openapi?format=JSON");
     }
 
     @Order(1)
@@ -82,4 +53,5 @@ public class OpenApiReactiveIT extends AbstractDbReactiveIT {
         Assertions.assertEquals(1, objectMap.keySet().size());
         Assertions.assertEquals(type, objectMap.keySet().iterator().next());
     }
+
 }
